[Bug] D125484: Selecting 'Reenter comments' when prompted by BK-TRIGGER-pre-commit-police-changeset-comments.rb only re-enters comments for the changeset, not for individual files in the changeset
src/e/scm/BitKeeper.java: If the deltas all succeeded but the commit was aborted, perhaps through comment policing, then the comment might be edited.  Change the comments on the files when we commit again, if we think there's nothing to delta, which would let us supply a new comment anyway.  This loses the sometimes-useful information that bk records as its default comment on bk cp and other such implicitly created deltas.

It also relies on the user doing Refresh List between the aborted Commit and the retry, in order for us to spot that our state is out of date, whereupon delta will fail with the like of:

2016-08-26T11:26:59.587-0700 CheckInTool: bk: stderr: delta: SCCS/s.anotherbadger.c is not locked.

(I think it doesn't return a failure, as we proceed with the commit and we do seem to check the status and throw if it's non-zero, so end up with a change set containing a file comment that's out of date.)

src/e/scm/CheckInWindow.java: That's not good, so run part of finish() even if commit() fails, in case it's partly succeeded.  I hope this won't automatically discard any of the user's check-in comment.  I don't think it will.  I worry about performance too but commit failure is a rare case and will usually be quick because of Linux's awesome file cache.