D131397: TC8884 errored with CommandNotFound("<class 'hnas_dev.command.dev_runner'> object HNASDev has no command 'simple_persistent_config'. Has the Bali CLI changed?",)
src/e/scm/Mercurial.java: I'm trying to follow the history of pythia, a hg repository, to find out why we restricted the Python HNAS CLI binding, generated by hnas/hnas/command.py's enumerate_available_commands(), to commands starting with a letter.  The file's been renamed several times.  hg log's --follow switch sorted the log and hg annotate is already going back through the renames.  It can tell me the filename at the time of the revision I first double-click on.  I use that to construct a followRenames method like the one I did in Git.java.  For my motivating use-case, I then shift-click to the previous revision to diff those.  Using the filename from the later revision gets me a diff in this particular case, but then I want to double-click on a red line to go to the code as it was in the earlier revision.  No line survives from that revision, so my method falls down.  Still, this helped a bit, so I'd like to Save Game.