src/e/scm/Git.java: git diff --name-status was showing the differences between the specified commit and the working copy.  That's not what listTouchedFilesInRevision is supposed to do.  Ryan McGeary supplies the right command in https://stackoverflow.com/a/424142/18096.
That still wasn't enough to get a non-empty Show Patch, though it did at least stop an exception being thrown.  We were reporting the previous revision as the same as the changed revision, meaning the difference between the two was inevitably empty.  Git has a special ^ syntax for getting the previous revision.  That's probably not going to do the right thing for merges, if there is a right way to give a patch for a three-way operation, but it's going to do something more useful than what we had.