Mercurial's hg commit without filenames, which we're compelled to use in the case of a merge, doesn't include new files but does include all changes to existing files.
src/e/scm/StatusesTableModel.java: So the test here needed to change...

src/e/scm/CheckInWindow.java:
src/e/scm/RevisionControlSystem.java: ... and the names would be misleading if I didn't rename them here...

src/e/scm/Mercurial.java: ... and the error message needs to say which files you must include and which you mustn't.  Well, surely we should include the right set for you and then not let you change the selection.  Ideally not by graying out the components, because that wouldn't let you discover the reason, but by throwing an error when you do.  But that'd be a bigger refactoring and might make it tricky to get the comment right, especially if you make an editing whoopsie.