D162605: gitlab merges messing up history
src/e/scm/Git.java: I've seen a couple of merge commits in a private gitlab instance that have a "body" that doesn't end in a newline.  That was messing up our parsing of the log output - we glommed the record for the next revision onto the end of that one and thus had no record of the next revision, causing exceptions on annotations, my struggling to understand the history and, embarrassingly, me to falsely accuse someone of blindly copy-and-pasting our LOG_SEPARATOR into a check-in comment.  I carefully remove the newline I add here, so that there's no change in what we pass onto the rest of the program, but now I'm sensitized to looking at the end of the comment field... why do all my comments end with a blank line?  I don't mind that I can put the caret on the line after the comment - I do end my comments, like my source files, with a newline.  There's a whole empty line after that, something I never deliberately do.

src/e/scm/Revision.java: That's because there was a newline appended here.  Move that from the end of the summary...

src/e/scm/RevisionView.java: ... to the comment separator, so we continue to always put the ASCII approximation of a horizontal rule on its own line, even for these unusual merge commits whose comments end with a gitlab merge request identifier.  Now the revisionCommentArea authentically preserves that lack of newline.  I like it when I can trust my tools to give me back what was put in.